requirements.txt
asyncpg==0.29.0
greenlet==3.0.3
pytest==8.0.2
pytest-asyncio==0.23.5
PyYAML==6.0.1
SQLAlchemy==2.0.27


iniconfig==2.0.0
packaging==23.2
pluggy==1.4.0
typing_extensions==4.9.0
